registration and login with axios and refactoring
